Updated the Webserver.js and index.html
Added a scrollbar for the chat window and added URLdecode for the name. Previously if you had a space in your name it would display as %20 but it has been fixed. Also added the scroll bar to the chat window so the window dosent get bigger if there is more then 26 messages.